refactor of structure
